requirements.txt
flask-diamond==0.4.5
bibtexparser==0.6.2
